requirements.txt
prettytable==3.8.0
fire==0.5.0
PyYAML==6.0.1
